Update discord, bot, main
