+++ main.py
@@ -1,4 +1,16 @@
-from src.bot import run_discord_bot
+from src.bot import BotDiscord
+import os
+from dotenv import load_dotenv
 
+load_dotenv()
 if __name__ =='__main__':
-    run_discord_bot()
+
+    bot = BotDiscord()
+
+
+    @bot.command()
+    async def chat_gpt(ctx, *args):
+        texte = ' '.join(args)
+        await bot.send_message(ctx, texte ,False)
+
+    bot.run(os.getenv("discord_token"))
